address Agis comments
